// apps/template/clients/admin/src/services/openapi/types.gen.ts
export type ApiTemplateV1GetMessageResponse = {
    id?: string;
    createdAt?: string;
    updatedAt?: string;
    title?: string;
    content?: string;
    status?: string;
};